chore: move kind helpers to internal.
