+++ packages/shared/common/src/internal/index.ts
@@ -2,3 +2,4 @@ export * from './diagnostics';
 export * from './evaluation';
 export * from './events';
 export * from './stream';
+export * from './context';
